+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/executor/ContributionsExecutor.java
@@ -1,9 +1,13 @@
 package org.heigit.ohsome.ohsomeapi.executor;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.SortedMap;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.heigit.ohsome.ohsomeapi.Application;
@@ -16,12 +20,19 @@ import org.heigit.ohsome.ohsomeapi.output.Description;
 import org.heigit.ohsome.ohsomeapi.output.Metadata;
 import org.heigit.ohsome.ohsomeapi.output.Response;
 import org.heigit.ohsome.ohsomeapi.output.contributions.ContributionsResult;
+import org.heigit.ohsome.ohsomeapi.output.groupby.GroupByResponse;
+import org.heigit.ohsome.ohsomeapi.output.groupby.GroupByResult;
+import org.heigit.ohsome.ohsomeapi.utils.GroupByBoundaryGeoJsonGenerator;
 import org.heigit.ohsome.oshdb.OSHDBTimestamp;
+import org.heigit.ohsome.oshdb.api.generic.OSHDBCombinedIndex;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer;
+import org.heigit.ohsome.oshdb.api.mapreducer.Mappable;
 import org.heigit.ohsome.oshdb.filter.FilterExpression;
 import org.heigit.ohsome.oshdb.util.celliterator.ContributionType;
+import org.heigit.ohsome.oshdb.util.function.SerializablePredicate;
 import org.heigit.ohsome.oshdb.util.mappable.OSMContribution;
 import org.locationtech.jts.geom.Geometry;
+import org.locationtech.jts.geom.Polygonal;
 
 /**
  * Includes the execute method for requests mapped to /contributions/count,
@@ -29,6 +40,7 @@ import org.locationtech.jts.geom.Geometry;
  * and /users/count.
  */
 public class ContributionsExecutor extends RequestExecutor {
+  private static final String CONTRIBUTION_TYPE_PARAMETER = "contributionType";
 
   private final InputProcessor inputProcessor;
   private final ProcessingData processingData;
@@ -148,24 +160,105 @@ public class ContributionsExecutor extends RequestExecutor {
       if (filter.isPresent()) {
         mapRedGroupByEntity = mapRedGroupByEntity.filter(filter.get());
       }
-      return contributionsFilter(mapRedGroupByEntity
-          .map(listContrsPerEntity -> listContrsPerEntity.get(listContrsPerEntity.size() - 1)))
-          .aggregateByTimestamp(OSMContribution::getTimestamp).count();
+      return mapRedGroupByEntity
+          .map(contributions -> contributions.get(contributions.size() - 1))
+          .filter(contributionsFilter(servletRequest.getParameter(CONTRIBUTION_TYPE_PARAMETER)))
+          .aggregateByTimestamp(OSMContribution::getTimestamp)
+          .count();
     } else {
-      return contributionsFilter(mapRed).aggregateByTimestamp().count();
+      return mapRed
+          .filter(contributionsFilter(servletRequest.getParameter(CONTRIBUTION_TYPE_PARAMETER)))
+          .aggregateByTimestamp()
+          .count();
     }
   }
 
   /**
-   * Filters contributions by contribution type.
+   * Handler method for count calculation of the endpoints
+   * /contributions/count/[density/]groupBy/boundary and /users/count/[density/]groupBy/boundary.
    *
-   * @param mapRed a MapReducer to be filtered
-   * @return MapReducer filtered by contribution type
+   * @return GroupByResponse {@link org.heigit.ohsome.ohsomeapi.output.Response Response}
+   * @throws Exception thrown by
+   *         {@link org.heigit.ohsome.ohsomeapi.inputprocessing.InputProcessor#processParameters()
+   *         processParameters},
+   *         {@link org.heigit.ohsome.ohsomeapi.inputprocessing.InputProcessor
+   *         #processParameters(ComputeMode) processParameters} and
+   *         {@link org.heigit.ohsome.oshdb.api.mapreducer.MapAggregator#count() count}
    */
-  private MapReducer<OSMContribution> contributionsFilter(MapReducer<OSMContribution> mapRed) {
-    String types = servletRequest.getParameter("contributionType");
+  public <P extends Geometry & Polygonal, V extends Comparable<V> & Serializable> Response
+      countGroupByBoundary(boolean isUsersRequest) throws Exception {
+    inputProcessor.getProcessingData().setGroupByBoundary(true);
+    var mapRed = inputProcessor.processParameters();
+    final var requestParameters = processingData.getRequestParameters();
+    List<Geometry> arrGeoms = processingData.getBoundaryList();
+    var arrGeomIds = inputProcessor.getUtils().getBoundaryIds();
+    @SuppressWarnings("unchecked")
+    // intentionally "unchecked" as check for P on Polygonal is already performed, and type of
+    // geomIds are either Strings or Integers which are both comparable and serializable
+    Map<V, P> geoms = IntStream.range(0, arrGeoms.size()).boxed()
+        .collect(Collectors.toMap(idx -> (V) arrGeomIds[idx], idx -> (P) arrGeoms.get(idx)));
+
+    var mapAgg = mapRed
+        .aggregateByTimestamp()
+        .aggregateByGeometry(geoms)
+        .map(OSMContribution.class::cast);
+
+    var filter = inputProcessor.getProcessingData().getFilterExpression();
+    if (filter.isPresent()) {
+      mapAgg = mapAgg.filter(filter.get());
+    }
+    if (isUsersRequest && processingData.isContainingSimpleFeatureTypes()) {
+      mapAgg = inputProcessor.filterOnSimpleFeatures(mapAgg);
+    }
+    SortedMap<OSHDBCombinedIndex<OSHDBTimestamp, V>, Integer> result;
+    if (isUsersRequest) {
+      result = mapAgg.map(OSMContribution::getContributorUserId).countUniq();
+    } else {
+      result = mapAgg
+          .filter(contributionsFilter(servletRequest.getParameter(CONTRIBUTION_TYPE_PARAMETER)))
+          .count();
+    }
+    var groupByResult = ExecutionUtils.nest(result);
+    var resultSet = groupByResult.entrySet().stream().map(entry ->
+        new GroupByResult(entry.getKey(), ExecutionUtils.fillContributionsResult(entry.getValue(),
+            requestParameters.isDensity(), inputProcessor, df, geoms.get(entry.getKey())
+    ))).toArray(GroupByResult[]::new);
+    Metadata metadata = null;
+    if (processingData.isShowMetadata()) {
+      long duration = System.currentTimeMillis() - startTime;
+      String description;
+      if (isUsersRequest) {
+        description = Description.countUsersGroupByBoundary(requestParameters.isDensity());
+      } else {
+        description = Description.countContributionsGroupByBoundary(requestParameters.isDensity());
+      }
+      metadata = new Metadata(duration, description,
+          inputProcessor.getRequestUrlIfGetRequest(servletRequest));
+    }
+    if ("geojson".equalsIgnoreCase(requestParameters.getFormat())) {
+      return GroupByResponse.of(new Attribution(URL, TEXT), Application.API_VERSION, metadata,
+          "FeatureCollection", GroupByBoundaryGeoJsonGenerator.createGeoJsonFeatures(resultSet,
+              processingData.getGeoJsonGeoms()));
+    } else if ("csv".equalsIgnoreCase(requestParameters.getFormat())) {
+      var exeUtils = new ExecutionUtils(processingData);
+      exeUtils.writeCsvResponse(resultSet, servletResponse,
+          ExecutionUtils.createCsvTopComments(URL, TEXT, Application.API_VERSION, metadata));
+      return null;
+    }
+    return new GroupByResponse(new Attribution(URL, TEXT), Application.API_VERSION, metadata,
+        resultSet);
+  }
+
+  /**
+   * Returns a function to filter contributions by contribution type.
+   *
+   * @param types the parameter string containing the to-be-filtered contribution types
+   * @return a lambda method implementing the filter which can be passed to
+   *         {@link Mappable#filter(SerializablePredicate)}
+   */
+  static SerializablePredicate<OSMContribution> contributionsFilter(String types) {
     if (types == null) {
-      return mapRed;
+      return ignored -> true;
     }
     types = types.toUpperCase();
     List<ContributionType> contributionTypes = new ArrayList<>();
@@ -188,6 +281,6 @@ public class ContributionsExecutor extends RequestExecutor {
               + "'geometryChange', 'tagChange' or a combination of them");
       }
     }
-    return mapRed.filter(contr -> contributionTypes.stream().anyMatch(contr::is));
+    return contr -> contributionTypes.stream().anyMatch(contr::is);
   }
 }